requirements.txt
discord.py
aiohttp
python-dotenv
python-Levenshtein
openai
